# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * South Asian Fashion E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductAvailability =
  (typeof ProductAvailability)[keyof typeof ProductAvailability];

export const ProductAvailability = {
  in_stock: "in_stock",
  made_to_order: "made_to_order",
  express: "express",
} as const;

export interface Product {
  id: number;
  slug: string;
  name: string;
  brand: string;
  sku: string;
  description: string;
  priceUsd?: number | null;
  pricePkr?: number | null;
  priceOnRequest: boolean;
  images: string[];
  fabric: string;
  pieces: number;
  colors: string[];
  embroidery: string;
  occasion: string;
  season: string;
  deliveryTime: string;
  availability: ProductAvailability;
  isNew: boolean;
  isFeatured: boolean;
  tags: string[];
  collection?: string | null;
  sizes: string[];
  createdAt: string;
}

export interface Collection {
  id: number;
  slug: string;
  name: string;
  description: string;
  bannerImage: string;
  productCount: number;
}

export interface Designer {
  id: number;
  name: string;
  slug: string;
  description: string;
  logoImage: string;
  coverImage: string;
  productCount: number;
}

export interface Review {
  id: number;
  customerName: string;
  location: string;
  rating: number;
  text: string;
  productName: string;
  avatarUrl?: string | null;
}

export interface CreateInquiryBody {
  name: string;
  email: string;
  phone?: string;
  dressInterest?: string;
  message: string;
}

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  dressInterest?: string | null;
  message: string;
  createdAt: string;
}

export type ListProductsParams = {
  occasion?: string;
  fabric?: string;
  color?: string;
  minPrice?: number;
  maxPrice?: number;
  designer?: string;
  collection?: string;
  season?: string;
  sort?: ListProductsSort;
  limit?: number;
  offset?: number;
};

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  newest: "newest",
  price_asc: "price_asc",
  price_desc: "price_desc",
  featured: "featured",
} as const;

export type ListProducts200 = {
  products: Product[];
  total: number;
};

export type GetFeaturedProducts200 = {
  products: Product[];
};

export type GetNewArrivalsParams = {
  limit?: number;
};

export type GetNewArrivals200 = {
  products: Product[];
};

export type GetRelatedProducts200 = {
  products: Product[];
};

export type ListCollections200 = {
  collections: Collection[];
};

export type GetCollection200 = {
  collection: Collection;
  products: Product[];
};

export type ListDesigners200 = {
  designers: Designer[];
};

export type ListReviews200 = {
  reviews: Review[];
};