# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * South Asian Fashion E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Collection {
  id: number;
  slug: string;
  name: string;
  description: string;
  bannerImage: string;
  productCount: number;
}