# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * South Asian Fashion E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductAvailability } from "./productAvailability";

export interface Product {
  id: number;
  slug: string;
  name: string;
  brand: string;
  sku: string;
  description: string;
  priceUsd?: number | null;
  pricePkr?: number | null;
  priceOnRequest: boolean;
  images: string[];
  fabric: string;
  pieces: number;
  colors: string[];
  embroidery: string;
  occasion: string;
  season: string;
  deliveryTime: string;
  availability: ProductAvailability;
  isNew: boolean;
  isFeatured: boolean;
  tags: string[];
  collection?: string | null;
  sizes: string[];
  createdAt: string;
}