# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * South Asian Fashion E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  customerName: string;
  location: string;
  rating: number;
  text: string;
  productName: string;
  avatarUrl?: string | null;
}